requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0rc0
google-auth==2.39.0
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.1.0
proto-plus==1.26.1
protobuf==4.21.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.6.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
tqdm==4.67.1
urllib3==2.4.0
Werkzeug==3.1.3
